Time: 2476 ms (14.25%), Space: 34.1 MB (99.15%) - LeetHub

+++ 0560-subarray-sum-equals-k/0560-subarray-sum-equals-k.cpp
@@ -4,11 +4,12 @@ public:
         int n = nums.size();
         int count = 0;
         for(int i=0;i<n;i++){
-            long long sum = 0;
+            int sum = 0;
             for(int j=i;j<n;j++){
                 sum += nums[j];
-                if(sum==k) {
+                if(sum == k){
                     count++;
+                    //break;
                 }
             }
         }
